src/end_to_end/cloud_server/message_processor/requirements.txt
paho_mqtt
tensorflow
opencv-python
numpy